[LCR-146] rename: dto명 수정
