+++ src/main/java/com/lgcns/domain/notification/service/NotificationService.java
@@ -1,9 +1,9 @@
 package com.lgcns.domain.notification.service;
 
-import com.lgcns.domain.notification.dto.NotificationResponseDTO;
+import com.lgcns.domain.notification.dto.response.NotificationResponse;
 import java.util.List;
 
 public interface NotificationService {
 
-    List<NotificationResponseDTO> findNotificationList(Long popupId);
+    List<NotificationResponse> findNotificationList(Long popupId);
 }
